Refactor(graduation.condition) : 졸업 조건을 사용하는 클래스 condition 분리 -> (credit, gpa, major, refinement)
