+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/credit/Credit.java
@@ -1,0 +1,7 @@
+package wiseSWlife.wiseSWlife.service.graduation.credit;
+
+import wiseSWlife.wiseSWlife.dto.graduation.form.CreditForm;
+
+public interface Credit {
+    CreditForm checkCredit(String sid, int myCredit);
+}
