+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/gpa/Gpa.java
@@ -1,0 +1,7 @@
+package wiseSWlife.wiseSWlife.service.graduation.gpa;
+
+import wiseSWlife.wiseSWlife.dto.graduation.form.GPAForm;
+
+public interface Gpa {
+    GPAForm checkGPA(String sid, double myGPA);
+}
